Unify postgres traits
