+++ apps/hash-graph/libs/graph/src/store/postgres/query/mod.rs
@@ -35,9 +35,18 @@ use crate::{
     subgraph::temporal_axes::QueryTemporalAxes,
 };
 
-pub trait PostgresRecord: Record {
+pub trait PostgresRecord: Record + QueryRecordDecode<Output = Self> {
+    type CompilationParameters: Send + 'static;
+
     /// The [`Table`] used for this `Query`.
     fn base_table() -> Table;
+
+    fn parameters() -> Self::CompilationParameters;
+
+    fn compile<'c, 'p: 'c>(
+        compiler: &mut SelectCompiler<'c, Self>,
+        paths: &'p Self::CompilationParameters,
+    ) -> Self::CompilationArtifacts;
 }
 
 /// An absolute path inside of a query pointing to an attribute.
@@ -81,17 +90,6 @@ pub trait PostgresSorting<R: Record>:
         compiler: &mut SelectCompiler<'c, R>,
         parameters: Option<&'c <Self::Cursor as QueryRecordEncode>::CompilationParameters<'p>>,
         temporal_axes: &QueryTemporalAxes,
-    ) -> Self::CompilationArtifacts;
-}
-
-pub trait QueryRecord: Record + QueryRecordDecode<Output = Self> {
-    type CompilationParameters: Send + 'static;
-
-    fn parameters() -> Self::CompilationParameters;
-
-    fn compile<'c, 'p: 'c>(
-        compiler: &mut SelectCompiler<'c, Self>,
-        paths: &'p Self::CompilationParameters,
     ) -> Self::CompilationArtifacts;
 }
